feat: implement react context on dark mode btn, & adjusting dark mode color in all pages
