+++ rock-paper-scissors-multiplayer/src/context/DarkModeContext.jsx
@@ -1,0 +1,31 @@
+import { createContext, useState, useEffect } from "react";
+
+export const DarkModeContext = createContext();
+
+export const DarkModeProvider = ({ children }) => {
+    // initiate state for dark mode
+    const [isDarkMode, setIsDarkMode] = useState(() => {
+        const savedTheme = localStorage.getItem("darkMode");
+        return savedTheme === "true";
+    });
+
+    const toggleDarkMode = () => {
+        setIsDarkMode(!isDarkMode);
+    }
+
+    useEffect(() => {
+        if (isDarkMode) {
+          document.documentElement.classList.add("dark");
+          localStorage.setItem("darkMode", "true");
+        } else {
+          document.documentElement.classList.remove("dark");
+          localStorage.setItem("darkMode", "false");
+        }
+    }, [isDarkMode]);
+
+    return (
+        <DarkModeContext.Provider value={{ isDarkMode, toggleDarkMode }}>
+          {children}
+        </DarkModeContext.Provider>
+    );
+};
