+++ rock-paper-scissors-multiplayer/src/main.jsx
@@ -1,10 +1,14 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './App.jsx'
+import { DarkModeProvider } from './context/DarkModeContext.jsx'
+import { BrowserRouter } from "react-router-dom";
 import './index.css'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <App />
+    <DarkModeProvider>
+      <App />
+    </DarkModeProvider>
   </StrictMode>,
 )
